Switch
Switch is now functional. I just need to insert the button to every page. I did my best in simplifying the CSS for light and dark, as well as cleaning up the javascript for the switch. Next time I will work on the button and click effects to make it presentable.